requirements.txt
Flask==1.1.2
gevent==20.6.2
gunicorn==20.0.4
opencv-python==4.4.0.42
tensorflow-cpu
Werkzeug==1.0.1
Pillow==7.0.0
